Feat: Add functionality to handle Mongoose Errors, rejections, and exceptions
Refactors the error-controller to handle errors related to Mongoose. Mongoose sends its own type of error object, the error controller refactors that error object into a more readable one. The server.js file is listening for the "unhandledRejection" and "uncaughtException" events to be thrown, and will handle shutting down the server. This acts a safety net for errors that aren't handled by the error-controller.