Feat: Add route protection to specific routes
Adds a protectRoute method in auth-controller. The protectRoute method check the request to see if a valid token was sent. The protectRoute method ensures that the client contains the proper credentials before moving onto the route handler. Adds a instance method to the UserSchema that checks if the current token is valid.