Feat: Add password reset functionality for users
Uses nodemailer to send an email with a reset token to the user who requested. The user can use the reset token, along with the new password, to reset their previous password. Set up mailtrap as the service/transporter. Adds forgotPassword and resetPassword routes. The forgotPassword handler will send the resetToken. The resetPassword handler will validate the token sent with the hashed token stored in the database.