Feat: Recreate Article Schema
Add additonal paths to the Article Schema. Set author path to not be required. Add the paths: introduction, tags, sections, and the virutal property tableOfContents. tableOfContents is created from the heading property in each objects sections array. The createArticle route is now protected and restricted to authors. The createArticle middleware uses the current logged in user's id, when creating the new article document.