Feat: Add file-upload middleware to handle saving images to server
Add new middleware for handling image upload to the server, named file-upload. The file-upload file exports a function called fileUpload. The fileUpload function is amulter middleware function with extra configuration. In article-routes, fileUpload middleware is used and expects a image property to be passed. Inside the createArticle function of article-controller, the image property of the Article model is set to the req.file.path.